requirements.txt
altair==4.2.2
attrs==22.2.0
bcrypt==4.0.1
blinker==1.5
boto3==1.26.74
botocore==1.29.74
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
entrypoints==0.4
extra-streamlit-components==0.1.56
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
PyMySQL==1.0.2
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.18.1
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
zipp==3.14.0
